requirements.txt
python-telegram-bot==13.7
requests==2.31.0
fastapi==0.109.2
python-multipart==0.0.9
uvicorn==0.27.1
psutil==5.9.8
Pillow==10.2.0
pytesseract==0.3.10
reportlab==4.0.8 